+++ examples/setup-temp/routes/index.tsx
@@ -1,26 +1,32 @@
 import { useSignal } from "@preact/signals";
-import { Main } from "../../../mod.ts";
+import { Layout, Main, Module } from "../../../mod.ts";
 import Counter from "../islands/Counter.tsx";
 
 export default function Home() {
   const count = useSignal(3);
   return (
-    <Main whitespace>
-      <div class="max-w-screen-md mx-auto flex flex-col items-center justify-center">
-        <img
-          class="my-6"
-          src="/logo.svg"
-          width="128"
-          height="128"
-          alt="the Fresh logo: a sliced lemon dripping with juice"
-        />
-        <h1 class="text-4xl font-bold">Welcome to Fresh</h1>
-        <p class="my-4">
-          Try updating this message in the
-          <code class="mx-2">./routes/index.tsx</code> file, and refresh.
-        </p>
-        <Counter count={count} />
-      </div>
-    </Main>
+    <>
+      <Main whitespace>
+        <Layout whitespace>
+          <Module size="full">
+            <div class="max-w-screen-md mx-auto flex flex-col items-center justify-center">
+              <img
+                class="my-6"
+                src="/logo.svg"
+                width="128"
+                height="128"
+                alt="the Fresh logo: a sliced lemon dripping with juice"
+              />
+              <h1 class="text-4xl font-bold">Welcome to Fresh</h1>
+              <p class="my-4">
+                Try updating this message in the
+                <code class="mx-2">./routes/index.tsx</code> file, and refresh.
+              </p>
+              <Counter count={count} />
+            </div>
+          </Module>
+        </Layout>
+      </Main>
+    </>
   );
 }