+++ src/styles.ts
@@ -1,23 +1,79 @@
 import { o } from './utils.ts';
 import { css } from '../deps.ts';
-import { iMain } from '../components/Main/setup.ts';
-import { iPanel } from '../components/Panel/setup.ts';
+import { iMain } from '../components/Main.tsx';
+import { iHeader } from '../components/Header.tsx';
+import { iLayout } from '../components/Layout.tsx';
+import { iModule } from '../components/Module.tsx';
 
 const clr = {
   bg_page: 'bg-page-50 dark:bg-d-page-50',
+  bg_panel: 'bg-panel dark:bg-d-panel',
 };
 
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 export const main = (props: iMain) => {
   props.class = o([
     clr.bg_page,
     'min-h-dvh',
     props.whitespace ? 'py-triple' : null,
   ], { ...props });
+
   return props;
 };
 
-export const panel = (props: iPanel) => {
-  props.class = o('bg-slate-100', { ...props });
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+export const header = (props: iHeader) => {
+  props.class = o([
+    clr.bg_panel,
+    'py-triple',
+    props.banner ? 'min-h-[90dvh]' : null,
+  ], { ...props });
+
+  return props;
+};
+
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+export const layout = (props: iLayout) => {
+  props.class = o(
+    [
+      'grid',
+      'grid-cols-6 md:grid-cols-12 lg:grid-cols-lg',
+      'gap-x-[0.8503100088rem] md:gap-x-[2.1257750221%] lg:gap-x-[1.5rem]',
+      'w-auto lg:w-min',
+      'mx-[0.8503100088rem] md:mx-[calc(23.13871222%-134.488008342px)] lg:mx-auto',
+    ],
+    { ...props },
+  );
+
+  return props;
+};
+
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+export const module = (props: iModule) => {
+  props.class = o(
+    [
+      // deno-fmt-ignore
+      props.half === 'none' ? 'hidden'
+      : props.half === 'full' ? 'col-span-12'
+      : props.half === 'xl' ? 'col-span-10'
+      : props.half === 'lg' ? 'col-span-8'
+      : props.half === 'md' ? 'col-span-6'
+      : props.half === 'sm' ? 'col-span-4'
+      : props.half === 'xs' ? 'col-span-2'
+      : null,
+      // deno-fmt-ignore
+      props.size === 'none' ? 'md:hidden'
+      : props.size === 'full' ? 'md:col-span-12'
+      : props.size === 'xl' ? 'md:col-span-10'
+      : props.size === 'lg' ? 'md:col-span-8'
+      : props.size === 'md' ? 'md:col-span-6'
+      : props.size === 'sm' ? 'md:col-span-4'
+      : props.size === 'xs' ? 'md:col-span-2'
+      : null,
+    ],
+    { ...props },
+  );
+
   return props;
 };
 
@@ -190,5 +246,64 @@ export const relativeContainer = css`
 
 position: relative;
 max-height: 0;
+
+`;
+
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+export const panelStyles = css`
+
+background-color: var(--clr-bg-panel);
+
+&.panel {
+
+  .button__style {
+    &--transparent {
+      background-color: var(--clr-bg-panel);
+
+      &:hover, &:focus {
+        background-color: var(--clr-bg-page-50);
+        border-bottom-color: var(--clr-txt-personality-10);
+      }
+
+      &:active {
+        background-color: var(--clr-txt-personality-10);
+        border-top-color: var(--clr-bg-personality-30);
+        border-bottom-color: transparent;
+      }
+    }
+
+    &--panel {
+      background-color: var(--clr-bg-page);
+
+      &:hover, &:focus {
+        background-color: var(--clr-txt-base-25);
+        border-bottom-color: var(--clr-txt-base);
+      }
+
+      &:active {
+        background-color: var(--clr-bg-page-50);
+        border-top-color: var(--clr-bg-personality-30);
+        border-bottom-color: transparent;
+      }
+    }
+  }
+
+  .code__wrapper {
+    background-color: var(--clr-txt-personality-10);
+  }
+
+  .kbd {
+    background-color: var(--clr-bg-page);
+    border-bottom-color: var(--clr-txt-personality-10);
+  }
+
+  .fieldset {
+    background-color: var(--clr-txt-base-10);
+
+    &:hover, &:has(:hover), &:has(:focus) {
+      background-color: var(--clr-txt-personality-10);
+    }
+  }
+}
 
 `;